Fix bugs related to partner registration

+++ src/main/java/oss/fosslight/service/impl/PartnerServiceImpl.java
@@ -182,9 +182,7 @@ public class PartnerServiceImpl extends CoTopComponent implements PartnerService
 	@Override
 	@Transactional
 	@CacheEvict(value="autocompletePartnerCache", allEntries=true)
-	public void registPartnerMaster(PartnerMaster partnerMaster, List<ProjectIdentification> ossComponents, List<List<ProjectIdentification>> ossComponentsLicense) {
-		PartnerMaster beforePartner =  partnerMapper.selectPartnerMaster(partnerMaster);
-		
+	public void registPartnerMaster(PartnerMaster partnerMaster, List<ProjectIdentification> ossComponents, List<List<ProjectIdentification>> ossComponentsLicense) {		
 		//파트너 마스터 테이블
 		if(partnerMaster.getPartnerId() != null) {
 			// admin이 아니라면 creator를 변경하지 않는다.
@@ -212,26 +210,30 @@ public class PartnerServiceImpl extends CoTopComponent implements PartnerService
 					T2File delFile = new T2File();
 					delFile.setFileSeq(fileSeq);
 					delFile.setGubn("A");
-					
-					fileMapper.updateFileDelYnKessan(delFile);
+
 					fileService.deletePhysicalFile(delFile, "PARTNER");
+					fileMapper.updateFileDelYnKessan(delFile);
 				}
 			}
 		}
 		
-		if(!isEmpty(beforePartner.getConfirmationFileId())) {
-			if(partnerMaster.getConfirmationFileId() == null || !partnerMaster.getConfirmationFileId().equals(beforePartner.getConfirmationFileId())) {
-				T2File delFile = new T2File();
-				delFile.setFileSeq(beforePartner.getConfirmationFileId());
-				fileService.deletePhysicalFile(delFile, "PARTNER");
-			}
-		}
+		PartnerMaster beforePartner =  partnerMapper.selectPartnerMaster(partnerMaster);
 		
-		if(!isEmpty(beforePartner.getOssFileId())) {
-			if(partnerMaster.getOssFileId() == null || !partnerMaster.getOssFileId().equals(beforePartner.getOssFileId())) {
-				T2File delFile = new T2File();
-				delFile.setFileSeq(beforePartner.getOssFileId());
-				fileService.deletePhysicalFile(delFile, "PARTNER");
+		if(beforePartner != null) {
+			if(!isEmpty(beforePartner.getConfirmationFileId())) {
+				if(partnerMaster.getConfirmationFileId() == null || !partnerMaster.getConfirmationFileId().equals(beforePartner.getConfirmationFileId())) {
+					T2File delFile = new T2File();
+					delFile.setFileSeq(beforePartner.getConfirmationFileId());
+					fileService.deletePhysicalFile(delFile, "PARTNER");
+				}
+			}
+			
+			if(!isEmpty(beforePartner.getOssFileId())) {
+				if(partnerMaster.getOssFileId() == null || !partnerMaster.getOssFileId().equals(beforePartner.getOssFileId())) {
+					T2File delFile = new T2File();
+					delFile.setFileSeq(beforePartner.getOssFileId());
+					fileService.deletePhysicalFile(delFile, "PARTNER");
+				}
 			}
 		}
